pyproject.toml
[project]
name = "sqlmesh-fabric"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
    "sqlmesh",
]

[tool.uv.sources]
sqlmesh = { git = "https://github.com/mattiasthalen/sqlmesh.git", rev = "add-support-for-fabric" }
